color-extract-service/requirements.txt
fastapi
uvicorn
pillow
numpy
scikit-learn
